+++ CHANGELOG.md
@@ -2,3 +2,4 @@
 - initial instructions and template for adding questions
 - Questions:
     - Reverse Integer
+    - Unique Paths